+++ Sources/Recycling/GridViews/Markets/GridView/MarketsGridView.swift
@@ -13,22 +13,11 @@ public protocol MarketsGridViewDataSource: class {
     func marketsGridView(_ marketsGridView: MarketsGridView, modelAtIndex index: Int) -> MarketsGridViewModel
 }
 
-private class MarketsFlowLayout: UICollectionViewFlowLayout {
-    override func invalidationContext(forBoundsChange newBounds: CGRect) -> UICollectionViewLayoutInvalidationContext {
-        let context = super.invalidationContext(forBoundsChange: newBounds)
-        if let context = context as? UICollectionViewFlowLayoutInvalidationContext, let collectionView = collectionView, !context.invalidateFlowLayoutDelegateMetrics {
-            let hasChangedWidth = newBounds.width != collectionView.bounds.width
-            context.invalidateFlowLayoutDelegateMetrics = hasChangedWidth
-        }
-        return context
-    }
-}
-
 public class MarketsGridView: UIView {
     // MARK: - Internal properties
 
     @objc private lazy var collectionView: UICollectionView = {
-        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: MarketsFlowLayout())
+        let collectionView = UICollectionView(frame: .zero, collectionViewLayout: MarketsGridViewFlowLayout())
         collectionView.delegate = self
         collectionView.dataSource = self
         collectionView.translatesAutoresizingMaskIntoConstraints = false
